requirements.txt
bcrypt==5.0.0
dnspython==2.8.0
mysql-connector-python==9.6.0
pillow==12.0.0
pymongo==4.16.0
python-dotenv==1.2.2
